+++ src/Http/Middleware/EmptyStringsToNull.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Rougin\Weasley\Http\Middleware;
+
+/**
+ * "Empty Strings To Null" Middleware
+ *
+ * @package Weasley
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class EmptyStringsToNull extends TransformRequest
+{
+    /**
+     * Transforms the specified value.
+     *
+     * @param  string $key
+     * @param  mixed  $value
+     * @return mixed
+     */
+    protected function transform($key, $value)
+    {
+        return is_string($value) && $value === '' ? null : $value;
+    }
+}