+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to `Weasley` will be documented in this file.
 ## [0.6.0](https://github.com/rougin/weasley/compare/v0.5.0...v0.6.0) - Unreleased
 
 ### Added
+- `EmptyStringsToNull` middleware
 - `Session` library
+- `TrimStrings` middleware
 
 ### Changed
 - `Generator` directory to `Console` directory